Add dependency to Projector module (fixes #1)
Migrations were failing when run during set up due to dependency
on `IRepository<MemberBindingRecord>`. This dependency is used to
allow a `Query` to be sorted by `Weight` property on `FeaturedPart`.
The `IRepository<MemberBindingRecord>` is made available by the
Projector module, thus a requirement for having it as a dependency.

Not sure if there is a better way of doing this yet, however whenever
we've had a featured mechanic in an Orchard site, we generally use
projections to display a list of feature content items.